+++ Dependencies/SupportModules/Il2Cpp/Main.cs
@@ -10,6 +10,7 @@ using UnityEngine;
 using Il2CppInterop.Common;
 using Microsoft.Extensions.Logging;
 using RedLoader.Fixes;
+using Color = System.Drawing.Color;
 using ILogger = Microsoft.Extensions.Logging.ILogger;
 
 namespace RedLoader.Support
@@ -33,7 +34,7 @@ namespace RedLoader.Support
                     InternalUtils.UnityInformationHandler.EngineVersion.Major,
                     InternalUtils.UnityInformationHandler.EngineVersion.Minor,
                     InternalUtils.UnityInformationHandler.EngineVersion.Build)
-            }).AddLogger(new InteropLogger())
+            }).AddLogger(new HarmonyExceptionLogger())
               .AddHarmonySupport();
             
             if (LaunchOptions.Console.CleanUnityLogs)
@@ -199,16 +200,24 @@ namespace RedLoader.Support
         public IDisposable BeginScope<TState>(TState state) => throw new NotImplementedException();
     }
     
-    public class AllLogger : ILogger
+    public class HarmonyExceptionLogger : ILogger
     {
+        private const string HarmonyMessage = "During invoking native->managed trampoline";
+            
         public void Log<TState>(LogLevel logLevel, EventId eventId, TState state, Exception exception, Func<TState, Exception, string> formatter)
         {
-            Console.WriteLine("ALLLOGGER: " + formatter(state, exception));
+            var msg = formatter(state, exception);
+            
+            if (exception != null && msg.Equals(HarmonyMessage))
+            {
+                RLog.MsgDirect(Color.Fuchsia, "====== HARMONY EXCEPTION ======");
+                RLog.MsgDirect(Color.IndianRed, exception.ToString());
+            }
         }
 
         public bool IsEnabled(LogLevel logLevel)
         {
-            return true;
+            return logLevel == LogLevel.Error;
         }
 
         public IDisposable BeginScope<TState>(TState state)